moved vulgar language to separate file and condensed check
